FIX LOG MISTERY
When a server was receiving partition information twice, this function
was being executed concurrently. Although we had a synchronous
datastructure, there was a race condition between the first and the
second if:
If both threads passed the outer if before adding the partition scheme
to the shared datastructures, when the tryAdd was run, then only one of
them would be able to add it. The other thread would raise an error to
the puppet master. By using simply the TryAdd, even if these threads are
adding concurrently, only one of them will add the value, but the other
won't throw any error